my commit3

+++ src/components/FormAuthIn.jsx
@@ -3,29 +3,52 @@ import FormAuth from "./FormAuth";
 import { AiOutlineArrowLeft, AiOutlineArrowRight } from "react-icons/ai";
 import { NavLink, useNavigate } from "react-router-dom";
 import { API_URL } from "../const2";
+import useLocalStorage from "use-local-storage";
 
 const FormAuthIn = () => {
   const navigate = useNavigate();
-  const onSubmit = (data) => {
-    // console.log(data);
-    axios.post(`${API_URL}/users`, {
-      email: data.email,
-      password: data.password,
-    });
-    localStorage.setItem(
-      "user",
-      JSON.stringify({
+
+  // const [user, setUser] = useLocalStorage("user", "");
+
+  const onSubmit = async (data) => {
+    try {
+      const isExistUser = await axios.get(
+        `${API_URL}/users?email=${data.email}`
+      );
+        const createdUserName = data.email.split("@")[0]
+
+      if (isExistUser.data.length > 0) {
+        alert("this user exists");
+        navigate(`../sign_up`);
+        return;
+      }
+
+      const res = await axios.post(`${API_URL}/users`, {
         email: data.email,
         password: data.password,
-      })
-    );
-    navigate(`../`)
-      .then((res) => console.log(res))
-      .catch((err) => console.log(err));
+        userName: createdUserName
+      });
+      localStorage.setItem(
+        "user",
+        JSON.stringify({
+          email: res.data.email,
+          password: res.data.password,
+          id: res.data.id,
+          userName: createdUserName
+        })
+      );
+      alert("You've registered and logged in");
+      navigate(`../`);
+    } catch (err) {
+      console.log(err);
+    }
   };
+
   return (
     <div className="bg-purple-100 h-screen">
-      <p className=" text-center text-2xl py-10 text-purple-950">Зарегистр</p>
+      <p className=" text-center text-2xl py-10 text-purple-950">
+      Create an account and log in
+      </p>
       <FormAuth onSubmit={onSubmit} />
       <div className=" mt-4 m-auto flex w-96 justify-between">
         <NavLink to={`../`}>
@@ -35,7 +58,7 @@ const FormAuthIn = () => {
         </NavLink>
         <NavLink to={`../sign_up`}>
           <p className=" text-right  text-blue-800">
-            Уже есть аккаунт <AiOutlineArrowRight className=" inline-block" />
+            Login <AiOutlineArrowRight className=" inline-block" />
           </p>
         </NavLink>
       </div>
